componetiza o multiselect distribuidor
